feat: cleanup, improved structure of microservice
