+++ console/api/dependencies/__init__.py
@@ -1,0 +1,14 @@
+# Copyright 2024 Piotr Łoboda.
+#
+# Use of this software is governed by the Business Source License
+# included in the file licenses/BSL.txt.
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the GNU Affero General Public License v3.0 only, included in the file
+# licenses/AGPL.txt.
+from .settings import settings
+from .meilisearch import meilisearch_client
+from .database import conn
+from .utils import parse_sql_update_query, camel_to_snake
+from .jwt import JWTBearer
